Add PostHog analytics to desktop main process

+++ surfsense_desktop/src/main.ts
@@ -11,6 +11,7 @@ import { registerQuickAsk, unregisterQuickAsk } from './modules/quick-ask';
 import { registerAutocomplete, unregisterAutocomplete } from './modules/autocomplete';
 import { registerIpcHandlers } from './ipc/handlers';
 import { createTray, destroyTray } from './modules/tray';
+import { initAnalytics, shutdownAnalytics, trackEvent } from './modules/analytics';
 
 registerGlobalErrorHandlers();
 
@@ -21,6 +22,8 @@ if (!setupDeepLinks()) {
 registerIpcHandlers();
 
 app.whenReady().then(async () => {
+  initAnalytics();
+  trackEvent('desktop_app_launched');
   setupMenu();
   try {
     await startNextServer();
@@ -69,9 +72,15 @@ app.on('before-quit', () => {
   isQuitting = true;
 });
 
-app.on('will-quit', () => {
+let didCleanup = false;
+app.on('will-quit', async (e) => {
+  if (didCleanup) return;
+  didCleanup = true;
+  e.preventDefault();
   unregisterQuickAsk();
   unregisterAutocomplete();
   unregisterFolderWatcher();
   destroyTray();
+  await shutdownAnalytics();
+  app.exit();
 });
